requirements.txt
alpaca-trade-api
flask
python-dotenv
pandas
